modify Class element
