+++ tltBlog/src/App.css
@@ -5,3 +5,15 @@
     @apply font-Qicksand
     }
 } ;
+
+@keyframes slideIn {
+  0% {
+    transform: translateX(-100%);
+  }
+  100% {
+    transform: translateX(0);
+  }
+}
+.animate-slide-in {
+  animation: slideIn 0.5s ease-in-out;
+}
