+++ tltBlog/src/pages/leftSide/leftSide.jsx
@@ -1,10 +1,9 @@
 import myAvatar from "../../images/myAvatar.jpg"; // di chuyển ra bên ngoài đường dẫn cùng cấp ../../..n/
 import myLogo from "../../images/myLogo.png"
 
-import PcCss from "./reponsive/PcCss";
-import MobileCss from "./reponsive/MobileCss";
+import PcCss from "./reponsiveLeftside/PcCss";
+import MobileCss from "./reponsiveLeftside/MobileCss";
 import NavBar from "./Navbar"
-import './reponsive/keyframe.css'
 
 import {AiOutlineMenuUnfold} from "react-icons/ai"
 
@@ -13,13 +12,13 @@ import React, { useState } from 'react';
 function LeftSife () { 
     return (
         <>
-            <div className={`xl:${PcCss.leftSide} s:${MobileCss.leftSideMobile}` }>
-                <div id="myProfile" className="font-PlayfairDisplay text-center flex flex-col items-center border-b-2">
-                    <img src={myAvatar} alt="" className={`drop-shadow-2xl w-36 h-36 rounded-full mx-16 mb-6 border-solid border-2 border-blue-500 cursor-pointer ${PcCss.hover}`}/>
-                    <h2 className="font-Kanit mb-3 text-xl font-bold ">Vũ Thắng</h2>
-                    <p className="mb-6 italic text-sm">"Design Your Life"</p>
+            <div className={`xl:${PcCss.leftSide} s:${MobileCss.leftSideMobile}`}>
+                <div id="myProfile" className={`${PcCss.profile.profileMain}`}>
+                    <img src={myAvatar} alt="" className={`${PcCss.profile.profileImg} ${PcCss.hover}`}/>
+                    <h2 className={`${PcCss.profile.profileName}`}>Vũ Thắng</h2>
+                    <p className={`${PcCss.profile.profileSologan}`}>"Design Your Life"</p>
                 </div>
-                <ul id="navBar" className="mt-6 font-Kanit font-bold text-sm text-textBlack text-center cursor-pointer">
+                <ul id="navBar" className={`${PcCss.navigation.navBar}`}>
                     <NavBar/>
                 </ul>
             </div>
