--- tltBlog/src/pages/leftSide/reponsive/MobileCss.jsx
@@ -1,7 +1,0 @@
-
-const MobileCss = {
-    leftSideMobile : "hidden",
-    // hover:"hover:border-b-2 hover:scale-x-125 hover:scale-y-125 transition-all duration-300 ease-in-out",
-}
-
-export default MobileCss
