--- tltBlog/src/pages/leftSide/reponsive/PcCss.jsx
@@ -1,7 +1,0 @@
-
-const PcCss = {
-    leftSide : "flex flex-col h-screen items-center bg-BgGrey py-11",
-    hover:"hover:border-b-2 hover:scale-x-125 hover:scale-y-125 transition-all duration-300 ease-in-out",
-}
-
-export default PcCss
