--- tltBlog/src/pages/leftSide/reponsive/keyframe.css
@@ -1,11 +1,0 @@
-@keyframes slideIn {
-  0% {
-    transform: translateX(-100%);
-  }
-  100% {
-    transform: translateX(0);
-  }
-}
-.animate-slide-in {
-  animation: slideIn 0.5s ease-in-out;
-}
