+++ tltBlog/src/pages/leftSide/reponsiveLeftside/MobileCss.jsx
@@ -1,0 +1,6 @@
+
+const MobileCss = {
+    leftSideMobile : "hidden",
+}
+
+export default MobileCss
